// frontend/src/api/generated/model/transactionOut.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Family Finance Tracker
 * OpenAPI spec version: 0.1.0
 */
import type { Money } from './money';
import type { TransactionOutTypeOverride } from './transactionOutTypeOverride';
import type { TransactionOutRawType } from './transactionOutRawType';
import type { TransactionOutCategoryRaw } from './transactionOutCategoryRaw';
import type { TransactionOutCategoryNormalized } from './transactionOutCategoryNormalized';
import type { TransactionOutCategoryOverride } from './transactionOutCategoryOverride';
import type { TransactionOutSubcategory } from './transactionOutSubcategory';
import type { TransactionOutOwnerId } from './transactionOutOwnerId';
import type { TransactionOutOwnerRaw } from './transactionOutOwnerRaw';
import type { TransactionOutMerchantRaw } from './transactionOutMerchantRaw';
import type { TransactionOutMerchantNormalized } from './transactionOutMerchantNormalized';
import type { TransactionOutMerchantOverride } from './transactionOutMerchantOverride';
import type { TransactionOutEffectiveType } from './transactionOutEffectiveType';
import type { TransactionOutEffectiveCategory } from './transactionOutEffectiveCategory';
import type { TransactionOutEffectiveMerchant } from './transactionOutEffectiveMerchant';

export interface TransactionOut {
  id: number;
  account_id: number;
  transaction_date: string;
  description: string;
  amount: Money;
  transaction_type: string;
  type_override?: TransactionOutTypeOverride;
  is_spend: boolean;
  raw_type?: TransactionOutRawType;
  category_raw: TransactionOutCategoryRaw;
  category_normalized: TransactionOutCategoryNormalized;
  category_override: TransactionOutCategoryOverride;
  subcategory?: TransactionOutSubcategory;
  owner_id: TransactionOutOwnerId;
  owner_raw?: TransactionOutOwnerRaw;
  merchant_raw?: TransactionOutMerchantRaw;
  merchant_normalized?: TransactionOutMerchantNormalized;
  merchant_override?: TransactionOutMerchantOverride;
  effective_type?: TransactionOutEffectiveType;
  effective_category?: TransactionOutEffectiveCategory;
  effective_merchant?: TransactionOutEffectiveMerchant;
}